Remove download current leading submission and put in submission.py (#193)
* Fix tests

* Remove test and fix cli
