+++ tests/test_submission.py
@@ -1,4 +1,6 @@
 """Test writeup functions"""
+import os
+import tempfile
 import unittest.mock as mock
 from unittest.mock import Mock, patch
 import uuid
@@ -7,13 +9,15 @@ import pytest
 import synapseclient
 from synapseclient.core.exceptions import SynapseHTTPError
 
-from challengeutils import dockertools, submission
+from challengeutils import dockertools, submission, utils
 
 SYN = mock.create_autospec(synapseclient.Synapse)
 SYN.configPath = None
 PROJ = mock.create_autospec(synapseclient.evaluation.Submission,
                             entityId="syn123",
                             entity=Mock(spec=synapseclient.Project))
+SUBMISSIONID = "111"
+QUEUEID = "333"
 
 
 # test that only Project entity type is accepted
@@ -176,3 +180,82 @@ def test_validate_docker_submission_notdocker():
          pytest.raises(ValueError,
                        match='Submission is not a Docker submission'):
         submission.validate_docker_submission(SYN, "123455")
+
+
+def test_nosub_get_submitterid_from_submission_id():
+    """Tests if submission id doesn't exist"""
+    with pytest.raises(ValueError,
+                       match=r'submission id*'),\
+        patch.object(utils, "evaluation_queue_query",
+                     return_value=[]) as patch_query:
+        submission.get_submitterid_from_submission_id(SYN, SUBMISSIONID,
+                                                      QUEUEID, verbose=False)
+        patch_query.assert_called_once()
+
+
+def test_get_submitterid_from_submission_id():
+    """Tests getting of submitter id"""
+    with patch.object(utils, "evaluation_queue_query",
+                      return_value=[{'submitterId': 1}]) as patch_query:
+        submitter = submission.get_submitterid_from_submission_id(
+            SYN, SUBMISSIONID, QUEUEID, verbose=False
+        )
+        patch_query.assert_called_once()
+        assert submitter == 1
+
+
+def test_get_submitters_lead_submission():
+    """Tests getting of lead submission"""
+    sub = synapseclient.Submission(evaluationId='2', entityId='2',
+                                   versionNumber='3')
+    temp = tempfile.NamedTemporaryFile()
+    sub.filePath = temp.name
+
+    with patch.object(utils, "evaluation_queue_query",
+                      return_value=[{'objectId': 1}]) as patch_query,\
+        patch.object(SYN, "getSubmission",
+                     return_value=sub) as patch_getsub:
+        dl_file = submission.get_submitters_lead_submission(
+            SYN, SUBMISSIONID, QUEUEID, "key", verbose=False
+        )
+        patch_query.assert_called_once()
+        patch_getsub.assert_called_once_with(1, downloadLocation='.')
+        assert dl_file == "previous_submission.csv"
+        os.unlink("previous_submission.csv")
+
+
+def test_none_get_submitters_lead_submission():
+    """Tests not getting submission"""
+    with patch.object(utils, "evaluation_queue_query",
+                      return_value=[]) as patch_query:
+        dl_file = submission.get_submitters_lead_submission(
+            SYN, SUBMISSIONID, QUEUEID, "key", verbose=False
+        )
+        patch_query.assert_called_once()
+        assert dl_file is None
+
+
+def test_download_current_lead_sub():
+    """Tests download of lead submission"""
+    sub = synapseclient.Submission(evaluationId='2', entityId='2',
+                                   versionNumber='3')
+    with patch.object(SYN, "getSubmission",
+                      return_value=sub) as patch_getsub,\
+        patch.object(submission, "get_submitterid_from_submission_id",
+                     return_value="2") as patch_getsubmitter,\
+        patch.object(submission, "get_submitters_lead_submission",
+                     return_value="path") as patch_get_lead:
+        dl_file = submission.download_current_lead_sub(
+            SYN, SUBMISSIONID, "VALIDATED", "key", verbose=False
+        )
+        patch_getsubmitter.assert_called_once()
+        patch_get_lead.assert_called_once()
+        assert dl_file == "path"
+
+
+def test_invalid_download_current_lead_sub():
+    """Tests None is downloaded if status is INVALID"""
+    dl_file = submission.download_current_lead_sub(
+        SYN, SUBMISSIONID, "INVALID", "key", verbose=False
+    )
+    assert dl_file is None
